requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
airtrain==0.1.68
altair==5.5.0
annotated-types==0.7.0
anthropic==0.51.0
anyio==4.9.0
appdirs==1.4.4
attrs==25.3.0
backoff==2.2.1
betterproto==2.0.0b7
black==25.1.0
blinker==1.9.0
boto3==1.38.12
botocore==1.38.12
CacheControl==0.14.3
cachetools==5.5.2
cerebras_cloud_sdk==1.29.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
dataclasses-json==0.6.7
datasets==3.6.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
eval_type_backport==0.2.2
faiss-cpu==1.11.0
filelock==3.18.0
firebase-admin==6.8.0
fireworks-ai==0.16.1
frozenlist==1.6.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-genai==1.14.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
groq==0.24.0
grpcio==1.71.0
grpcio-status==1.71.0
grpclib==0.4.8
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
httpx-ws==0.7.2
huggingface-hub==0.31.2
hyperframe==6.1.0
idna==3.10
isort==5.13.2
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
langchain==0.3.25
langchain-community==0.3.23
langchain-core==0.3.59
langchain-openai==0.3.16
langchain-text-splitters==0.3.8
langgraph==0.4.3
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.66
langsmith==0.3.42
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.8.0
mdurl==0.1.2
modelcontextprotocol==0.1.0
msgpack==1.1.0
multidict==6.4.3
multiprocess==0.70.16
mypy_extensions==1.1.0
narwhals==1.38.2
nest-asyncio==1.6.0
numpy==2.2.5
openai==1.78.0
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
posthog==4.0.1
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
ragas==0.2.15
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
s3transfer==0.12.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
sse-starlette==2.3.4
starlette==0.46.2
streamlit==1.45.0
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.9.0
together==1.5.7
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typer==0.15.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
wcwidth==0.2.13
websockets==15.0.1
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0